Split activeDevicesGraph
